Refactor common code to abstract PowerSource class
