+++ src/main/java/oshi/hardware/platform/mac/MacPowerSource.java
@@ -19,10 +19,6 @@ package oshi.hardware.platform.mac;
 import java.util.ArrayList;
 import java.util.List;
 
-import javax.json.Json;
-import javax.json.JsonBuilderFactory;
-import javax.json.JsonObject;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -31,11 +27,11 @@ import com.sun.jna.Pointer;
 import com.sun.jna.ptr.IntByReference;
 
 import oshi.hardware.PowerSource;
+import oshi.hardware.common.AbstractPowerSource;
 import oshi.jna.platform.mac.CoreFoundation;
 import oshi.jna.platform.mac.CoreFoundation.CFArrayRef;
 import oshi.jna.platform.mac.CoreFoundation.CFDictionaryRef;
 import oshi.jna.platform.mac.CoreFoundation.CFTypeRef;
-import oshi.json.NullAwareJsonObjectBuilder;
 import oshi.jna.platform.mac.IOKit;
 
 /**
@@ -43,43 +39,12 @@ import oshi.jna.platform.mac.IOKit;
  * 
  * @author widdis[at]gmail[dot]com
  */
-public class MacPowerSource implements PowerSource {
+public class MacPowerSource extends AbstractPowerSource {
     private static final Logger LOG = LoggerFactory.getLogger(MacPowerSource.class);
 
-    private String name;
-
-    private double remainingCapacity;
-
-    private double timeRemaining;
-
-    private JsonBuilderFactory jsonFactory = Json.createBuilderFactory(null);
-
     public MacPowerSource(String newName, double newRemainingCapacity, double newTimeRemaining) {
-        this.name = newName;
-        this.remainingCapacity = newRemainingCapacity;
-        this.timeRemaining = newTimeRemaining;
+        super(newName, newRemainingCapacity, newTimeRemaining);
         LOG.debug("Initialized MacPowerSource");
-    }
-
-    @Override
-    public String getName() {
-        return this.name;
-    }
-
-    @Override
-    public double getRemainingCapacity() {
-        return this.remainingCapacity;
-    }
-
-    @Override
-    public double getTimeRemaining() {
-        return this.timeRemaining;
-    }
-
-    @Override
-    public JsonObject toJSON() {
-        return NullAwareJsonObjectBuilder.wrap(jsonFactory.createObjectBuilder()).add("name", getName()).add("remainingCapacity", getRemainingCapacity())
-                .add("timeRemaining", getTimeRemaining()).build();
     }
 
     /**
